+++ docs/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.21] - 2026-07-26
+
+### Added
+
+- Command `playlist download`, clearing the queue, playing the given playlist, and downloading
+  the resulting queue like `queue download`, with the options of both commands
+- Subsection for the `playlist download` subcommand of the `downloads` section
+  of the configuration file
+
+
 ## [0.0.20] - 2026-07-26
 
 ### Added